Move base menu classes to separate package
