+++ src/main/java/com/verdantartifice/primalmagick/common/menus/InfernalFurnaceMenu.java
@@ -2,6 +2,8 @@ package com.verdantartifice.primalmagick.common.menus;
 
 import com.verdantartifice.primalmagick.PrimalMagick;
 import com.verdantartifice.primalmagick.common.crafting.recipe_book.ArcaneRecipeBookType;
+import com.verdantartifice.primalmagick.common.menus.base.AbstractTileInventoryMenu;
+import com.verdantartifice.primalmagick.common.menus.base.IArcaneRecipeBookMenu;
 import com.verdantartifice.primalmagick.common.menus.slots.FilteredSlot;
 import com.verdantartifice.primalmagick.common.menus.slots.InfernalFurnaceResultSlot;
 import com.verdantartifice.primalmagick.common.menus.slots.WandSlot;
