+++ src/main/java/com/verdantartifice/primalmagick/common/menus/base/AbstractTileMenu.java
@@ -1,4 +1,4 @@
-package com.verdantartifice.primalmagick.common.menus;
+package com.verdantartifice.primalmagick.common.menus.base;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
